requirements.txt
torch==2.1.0
torchvision==0.16.0
ftfy==6.1.1
regex==2023.8.8
faiss-cpu==1.7.4
Pillow==10.0.0
numpy==1.24.3
tqdm==4.65.0
flask==2.3.3
flask-cors==4.0.0